Sound Manager and Sliding Door
Added the sliding door which can be tested by pressing l (open) and k
(close). the test code is in the update function of slidingDoor2, and
should be removed after alissia pulls the trigger from it (or whenever
we are done testing it).
ALSO, I added a basic implementation of a sound test (press j to hear a
localized sound). I included basic instructions for setting up
additional sounds within the slidingDoor2 script as well:
/*example of playing a sound. doesnt work well with update calling the
sound multiple times (and im just trying to show proof of concept)
* in the code where it will actually be used, it should only called once
per sound. import sound files into the AudioClips file
* to add additional sounds, copy an existing prefab within SoundGroups
and modify the elements by including the sound pool you want to play
* pitch is optional. make sure spatial blend is set to 3d. sound prefabs
must remain in SoundGroups folder to be located by the script
* create a variable for the object creating the sound <doorSFX> and in
the inspector set that field to the same name as the prefab you want
from SoundGroups
* some additional sounds are in the googledrive*/

additionally, there is a sound test scene that came with the aduio
manager if you want to do any further testing.